refactor(server): move genmodemask to gamemode.hpp
TODO: this is just refactoring, but it should be simply deleted.
Its actually a parser for the arguments for the `maprotation` and `teamkillkick` command flags (both will go into InexorFlex)
